Revert "store CGM and pump state in binary plist files instead of UserDefault…"
