+++ FreeAPS/Sources/APS/Extensions/UserDefaultsExtensions.swift
@@ -5,29 +5,20 @@ import RileyLinkKit
 
 extension UserDefaults {
     private enum Key: String {
-        case legacyPumpManagerState = "com.rileylink.PumpManagerRawValue"
-        case legacyCGMManagerState = "cgmManagerRawValue"
+        case pumpManagerRawValue = "com.rileylink.PumpManagerRawValue"
     }
 
-    var legacyPumpManagerRawValue: PumpManager.RawStateValue? {
+    var pumpManagerRawValue: PumpManager.RawStateValue? {
         get {
-            dictionary(forKey: Key.legacyPumpManagerState.rawValue)
+            dictionary(forKey: Key.pumpManagerRawValue.rawValue)
         }
         set {
-            set(newValue, forKey: Key.legacyPumpManagerState.rawValue)
+            set(newValue, forKey: Key.pumpManagerRawValue.rawValue)
         }
     }
 
-    func clearLegacyPumpManagerRawValue() {
-        set(nil, forKey: Key.legacyPumpManagerState.rawValue)
-    }
-
-    var legacyCgmManagerRawValue: CGMManager.RawStateValue? {
-        get { object(forKey: Key.legacyCGMManagerState.rawValue) as? CGMManager.RawStateValue }
-        set { set(newValue, forKey: Key.legacyCGMManagerState.rawValue) }
-    }
-
-    func clearLegacyCGMManagerRawValue() {
-        set(nil, forKey: Key.legacyCGMManagerState.rawValue)
+    var cgmManagerRawValue: CGMManager.RawStateValue? {
+        get { object(forKey: "cgmManagerRawValue") as? CGMManager.RawStateValue }
+        set { set(newValue, forKey: "cgmManagerRawValue") }
     }
 }
